+++ src/.vuepress/components/FixtureInfo/FixtureGroups.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import * as fixtureConfigUnknown from "../../../lights/fixtureConfig.json";
+import * as fixtureConfigUnknown from "../../../data/gen/kistanFixtures.json";
 import { FixtureConfig, FixtureGroupSchema } from './fixtureConfigSchema';
 import FixtureGroup from "./FixtureGroup.vue";
 const fixtureConfig = fixtureConfigUnknown as FixtureConfig;
